# fe/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Economy Cloud Data API",
    "version": "1.1.0",
    "description": "Authenticated optional persistence for a minimal profile, versioned cloud consent and a privacy-minimised snapshot. Local use does not call this API. Accept the current consent statement before testing snapshot writes."
  },
  "servers": [
    { "url": "/", "description": "Current origin" }
  ],
...
  "paths": {
...
    "/api/snapshots": {
      "get": {
        "tags": ["Snapshots"],
        "summary": "Load the current user's snapshot",
        "operationId": "loadSnapshot",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Snapshot found, or null when none is stored",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SnapshotResponse" } } }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/CloudFailure" },
          "503": { "$ref": "#/components/responses/NotConfigured" }
        }
      },
      "put": {
        "tags": ["Snapshots"],
        "summary": "Create or replace the current user's snapshot",
        "operationId": "saveSnapshot",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SnapshotInput" },
              "example": {
                "schemaVersion": 2,
                "payload": { "tx": [], "overrides": {}, "rules": [], "cats": [], "budgets": {} }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored snapshot",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SnapshotResponse" } } }
          },
          "400": { "$ref": "#/components/responses/InvalidSnapshot" },
          "413": { "$ref": "#/components/responses/TooLarge" },
          "415": { "$ref": "#/components/responses/JsonRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "403": { "$ref": "#/components/responses/ConsentRequired" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/CloudFailure" },
          "503": { "$ref": "#/components/responses/NotConfigured" }
        }
      },
      "delete": {
        "tags": ["Snapshots"],
        "summary": "Delete the current user's cloud snapshot",
        "operationId": "deleteSnapshot",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Snapshot deleted" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/CloudFailure" },
          "503": { "$ref": "#/components/responses/NotConfigured" }
        }
      }
    },
    "/api/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Load the current user's preferred locale",
        "operationId": "loadProfile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Profile found, or null", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileResponse" } } } },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/CloudFailure" },
          "503": { "$ref": "#/components/responses/NotConfigured" }
        }
      },
      "put": {
        "tags": ["Profile"],
        "summary": "Create or update the current user's preferred locale",
        "operationId": "saveProfile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileInput" } } } },
        "responses": {
          "200": { "description": "Stored profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileResponse" } } } },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "413": { "$ref": "#/components/responses/RequestTooLarge" },
          "415": { "$ref": "#/components/responses/JsonRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/CloudFailure" },
          "503": { "$ref": "#/components/responses/NotConfigured" }
        }
      }
    },
    "/api/consents/cloud-sync": {
      "get": {
        "tags": ["Consent"],
        "summary": "Load the current cloud-sync consent statement",
        "operationId": "loadCloudConsent",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Current statement record, or null", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConsentResponse" } } } },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/CloudFailure" },
          "503": { "$ref": "#/components/responses/NotConfigured" }
        }
      },
      "put": {
        "tags": ["Consent"],
        "summary": "Accept the current cloud-sync consent statement",
        "operationId": "acceptCloudConsent",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConsentInput" } } } },
        "responses": {
          "200": { "description": "Stored consent", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConsentResponse" } } } },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "413": { "$ref": "#/components/responses/RequestTooLarge" },
          "415": { "$ref": "#/components/responses/JsonRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/CloudFailure" },
          "503": { "$ref": "#/components/responses/NotConfigured" }
        }
      },
      "delete": {
        "tags": ["Consent"],
        "summary": "Withdraw the current cloud-sync consent statement",
        "operationId": "withdrawCloudConsent",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Consent withdrawn, or no current record existed" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/CloudFailure" },
          "503": { "$ref": "#/components/responses/NotConfigured" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "CloudStatePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["tx", "overrides", "rules", "cats", "budgets"],
        "properties": {
          "tx": { "type": "array", "maxItems": 50000, "items": { "$ref": "#/components/schemas/PrivacySafeTransaction" } },
          "overrides": { "type": "object", "maxProperties": 50000, "additionalProperties": { "type": "string", "maxLength": 200 } },
          "rules": { "type": "array", "maxItems": 1000, "items": { "$ref": "#/components/schemas/CategoryRule" } },
          "cats": { "type": "array", "maxItems": 1000, "items": { "$ref": "#/components/schemas/Category" } },
          "budgets": { "type": "object", "maxProperties": 1000, "additionalProperties": { "type": "number", "minimum": 0 } },
          "goals": {
            "type": "array", "maxItems": 100,
            "items": {
              "type": "object", "additionalProperties": false,
              "required": ["id", "name", "target", "saved"],
              "properties": {
                "id": { "type": "string", "maxLength": 100 },
                "name": { "type": "string", "maxLength": 200 },
                "target": { "type": "number", "minimum": 0, "maximum": 1000000000 },
                "saved": { "type": "number", "minimum": 0, "maximum": 1000000000 },
                "due": { "type": ["string", "null"], "maxLength": 7, "pattern": "^\\d{4}-\\d{2}$" }
              }
            }
          }
        }
      },
      "PrivacySafeTransaction": {
        "type": "object",
        "description": "Minimised transaction. Account/card identifiers, CVV, original references and filenames are excluded.",
        "required": ["date", "vdate", "ref", "desc", "out", "in", "bal", "pending", "src"],
        "properties": {
          "date": { "type": "string", "maxLength": 32 },
          "vdate": { "type": "string", "maxLength": 32 },
          "ref": { "type": "string", "const": "" },
          "src": { "type": "string", "enum": ["bank-report", "card-report", "manual-entry", "open-banking"] },
          "desc": { "type": "string", "maxLength": 500 },
          "out": { "type": "number" },
          "in": { "type": "number" },
          "bal": { "oneOf": [{ "type": "number" }, { "type": "null" }] },
          "pending": { "type": "boolean" },
          "source": { "type": "string", "enum": ["bank", "card"] },
          "cardKind": { "type": "string", "enum": ["bank", "external"], "description": "Who issues the card. Decides whether its detail cancels the statement settlement." },
          "cardBrand": { "type": "string", "enum": ["visa", "cal", "isracard", "diners", "amex", "max", "leumi", "other"], "description": "Which company issued the card, as the customer answered at import. Provenance only." },
              "cardLast4": { "type": ["string", "null"], "pattern": "^\\d{4}$" },
          "id": { "type": "string", "maxLength": 200 },
          "cat": { "type": "string", "maxLength": 100 },
          "kind": { "type": "string", "enum": ["expense", "income", "neutral"] }
        },
        "additionalProperties": false
      },
      "CategoryRule": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "match", "cat"],
        "properties": {
          "id": { "type": "string", "maxLength": 100 },
          "match": { "type": "string", "maxLength": 200 },
          "cat": { "type": "string", "maxLength": 100 },
          "when": { "type": "string", "enum": ["in", "out"], "description": "Direction the rule reads, for wording that means opposite things on each side of a statement." }
        }
      },
      "Category": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "kind"],
        "properties": {
          "id": { "type": "string", "maxLength": 100 },
          "name": { "type": "string", "maxLength": 200 },
          "kind": { "type": "string", "enum": ["expense", "income", "neutral"] }
        }
      },
      "SnapshotInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schemaVersion", "payload"],
        "properties": {
          "schemaVersion": { "type": "integer", "const": 2 },
          "payload": { "$ref": "#/components/schemas/CloudStatePayload" }
        }
      },
      "Snapshot": {
        "allOf": [
          { "$ref": "#/components/schemas/SnapshotInput" },
          {
            "type": "object",
            "properties": { "updatedAt": { "type": "string", "format": "date-time" } }
          }
        ]
      },
      "SnapshotResponse": {
        "type": "object",
        "required": ["snapshot"],
        "properties": { "snapshot": { "oneOf": [{ "$ref": "#/components/schemas/Snapshot" }, { "type": "null" }] } }
      },
      "ProfileInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["preferredLocale"],
        "properties": { "preferredLocale": { "type": "string", "enum": ["he", "en", "am", "fr"] } }
      },
      "Profile": {
        "allOf": [
          { "$ref": "#/components/schemas/ProfileInput" },
          { "type": "object", "required": ["createdAt", "updatedAt"], "properties": {
            "createdAt": { "type": "string", "format": "date-time" },
            "updatedAt": { "type": "string", "format": "date-time" }
          } }
        ]
      },
      "ProfileResponse": {
        "type": "object",
        "required": ["profile"],
        "properties": { "profile": { "oneOf": [{ "$ref": "#/components/schemas/Profile" }, { "type": "null" }] } }
      },
      "ConsentInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["locale"],
        "properties": { "locale": { "type": "string", "enum": ["he", "en", "am", "fr"] } }
      },
      "Consent": {
        "type": "object",
        "additionalProperties": false,
        "required": ["purpose", "statementVersion", "locale", "acceptedAt", "withdrawnAt"],
        "properties": {
          "purpose": { "type": "string", "const": "cloud_sync" },
          "statementVersion": { "type": "string", "const": "cloud-sync-v2-privacy-minimised-2026-08-24" },
          "locale": { "type": "string", "enum": ["he", "en", "am", "fr"] },
          "acceptedAt": { "type": "string", "format": "date-time" },
          "withdrawnAt": { "oneOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }] }
        }
      },
      "ConsentResponse": {
        "type": "object",
        "required": ["consent"],
        "properties": { "consent": { "oneOf": [{ "$ref": "#/components/schemas/Consent" }, { "type": "null" }] } }
      },
      "Error": {
        "type": "object",
        "required": ["code"],
        "properties": { "code": { "type": "string" } }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid user JWT",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ConsentRequired": {
        "description": "The current cloud-sync statement must be actively accepted before a snapshot write",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "InvalidRequest": {
        "description": "The request body is invalid or unsupported",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "InvalidSnapshot": {
        "description": "Invalid, unsupported or oversized snapshot",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "TooLarge": {
        "description": "The snapshot exceeds the request or payload limit",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "RequestTooLarge": {
        "description": "The metadata request exceeds the 1 KiB limit",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "JsonRequired": {
        "description": "PUT requests require application/json",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "RateLimited": {
        "description": "Too many requests from the same client window",
        "headers": { "Retry-After": { "schema": { "type": "integer" } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "CloudFailure": {
        "description": "Supabase authentication or persistence failed without exposing provider details",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotConfigured": {
        "description": "Supabase environment variables are not configured",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  }
}